requirements.txt
django
django-extensions
django-model-utils
Faker